pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tabcan"
version = "0.1.0"
description = "TabCAN: Tabular Causal Attention Network"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "MOSTLY AI" }]
keywords = ["transformer", "tabular", "machine-learning", "synthetic-data", "autoregressive", "attention"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.0.0",
    "hf_transfer",
    "pyarrow",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
benchmark = [
    "mostlyai-qa>=0.1.0",
    "seaborn",
]

[project.urls]
Homepage = "https://github.com/mostly-ai/tabcan"
Repository = "https://github.com/mostly-ai/tabcan"

[tool.hatch.build.targets.wheel]
packages = ["src/tabcan"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
